requirements.txt
numpy
matplotlib
h5py
pynrrd
myst-parser
coverage